feat: Added api for favorites component
Added api for favorites component and made  a common module for Home and Favorites component for
rendering the cards.
